+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Fixed
+- Microsecond logging for nested steps, by @HardNorth
 
 ## [5.7.3]
 ### Added